move most code into lib.rs to keep main.rs clean
